requirements.txt
Flask==1.1.2
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
jedi==0.17.2
Flask-WTF==0.14.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1